Avoid duplicate object names by adding a suffix to armature names
In clump pages, the clump and other chunks can have the same name, as long as their types are different.

In blender, an armature and its child mesh cannot have the same name, so we add a suffix (that is removed when exporting, if it exists) to avoid blender renaming objects.